+++ src/js/tools/metroconfig.ts
@@ -1,0 +1,77 @@
+import type { MetroConfig } from 'metro';
+
+import { createSentryMetroSerializer } from './sentryMetroSerializer';
+
+export * from './sentryMetroSerializer';
+
+/**
+ * Adds Sentry to the Metro config.
+ *
+ * Adds Debug ID to the output bundle and source maps.
+ * Collapses Sentry frames from the stack trace view in LogBox.
+ */
+export function withSentry(config: MetroConfig): MetroConfig {
+  let newConfig = config;
+
+  newConfig = withSentryDebugId(newConfig);
+  newConfig = withSentryFramesCollapsed(newConfig);
+
+  return newConfig;
+}
+
+type MetroCustomSerializer = Required<Required<MetroConfig>['serializer']>['customSerializer'] | undefined;
+
+function withSentryDebugId(config: MetroConfig): MetroConfig {
+  const customSerializer = createSentryMetroSerializer(
+    config.serializer?.customSerializer || undefined,
+  ) as MetroCustomSerializer;
+  // MetroConfig types customSerializers as async only, but sync returns are also supported
+  // The default serializer is sync
+
+  return {
+    ...config,
+    serializer: {
+      ...config.serializer,
+      customSerializer,
+    },
+  };
+}
+
+type MetroFrame = Parameters<Required<Required<MetroConfig>['symbolicator']>['customizeFrame']>[0];
+type MetroCustomizeFrame = { readonly collapse?: boolean };
+type MetroCustomizeFrameReturnValue =
+  | ReturnType<Required<Required<MetroConfig>['symbolicator']>['customizeFrame']>
+  | undefined;
+
+function withSentryFramesCollapsed(config: MetroConfig): MetroConfig {
+  const originalCustomizeFrame = config.symbolicator?.customizeFrame;
+  const collapseSentryInternalFrames = (frame: MetroFrame): boolean =>
+    typeof frame.file === 'string' && frame.file.includes('node_modules/@sentry/utils/cjs/instrument.js');
+
+  const customizeFrame = (frame: MetroFrame): MetroCustomizeFrameReturnValue => {
+    const originalOrSentryCustomizeFrame = (
+      originalCustomization: MetroCustomizeFrame | undefined,
+    ): MetroCustomizeFrame => ({
+      ...originalCustomization,
+      collapse: (originalCustomization && originalCustomization.collapse) || collapseSentryInternalFrames(frame),
+    });
+
+    const maybePromiseCustomization = (originalCustomizeFrame && originalCustomizeFrame(frame)) || undefined;
+
+    if (maybePromiseCustomization !== undefined && 'then' in maybePromiseCustomization) {
+      return maybePromiseCustomization.then<MetroCustomizeFrame>(originalCustomization =>
+        originalOrSentryCustomizeFrame(originalCustomization),
+      );
+    }
+
+    return originalOrSentryCustomizeFrame(maybePromiseCustomization);
+  };
+
+  return {
+    ...config,
+    symbolicator: {
+      ...config.symbolicator,
+      customizeFrame,
+    },
+  };
+}